Holy shit, I will update the changelog tomorrow

+++ Frosty Cheeks/Frosty Cheeks/Meter.cs
@@ -13,14 +13,17 @@ namespace Frosty_Cheeks
 {
     class Meter:GUIObject
     {
-        private int coldMeter; // how cold they are, blue meter
-        public int ColdMeter
+        private float coldMeter; // how cold they are, blue meter
+        public float ColdMeter
         {
             get { return coldMeter; }
-            set { coldMeter = value; }
+            set 
+            {
+                coldMeter = value;
+            }
         }
-        private const int METER_MAX = 100;
-        public int Meter_Max
+        private const float METER_MAX = 100;
+        public float Meter_Max
         {
             get { return METER_MAX; }
         }
